backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
sqlalchemy==2.0.19
aiomysql==0.1.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.11.1
pydantic==2.1.1
python-dotenv==1.0.0